Move some classes into new packages Add packages: - ui, ui.adapters, ui.fragments - repository
