+++ app/src/main/java/ru/aevd/androidacademymovieapp/MainActivity.kt
@@ -3,6 +3,8 @@ package ru.aevd.androidacademymovieapp
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import ru.aevd.androidacademymovieapp.data.Movie
+import ru.aevd.androidacademymovieapp.ui.fragments.FragmentMoviesDetails
+import ru.aevd.androidacademymovieapp.ui.fragments.FragmentMoviesList
 
 class MainActivity: AppCompatActivity(), TransactionsFragmentClicks {
     private val moviesListFragment: FragmentMoviesList = FragmentMoviesList()
