+++ app/src/main/java/ru/aevd/androidacademymovieapp/repository/GetMoviesUseCase.kt
@@ -1,4 +1,4 @@
-package ru.aevd.androidacademymovieapp
+package ru.aevd.androidacademymovieapp.repository
 
 import ru.aevd.androidacademymovieapp.data.Movie
 import ru.aevd.androidacademymovieapp.network.NetworkOperations
