+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/adapters/ActorsAdapter.kt
@@ -1,4 +1,4 @@
-package ru.aevd.androidacademymovieapp
+package ru.aevd.androidacademymovieapp.ui.adapters
 
 import android.view.LayoutInflater
 import android.view.View
@@ -10,6 +10,7 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.bumptech.glide.request.RequestOptions
+import ru.aevd.androidacademymovieapp.R
 import ru.aevd.androidacademymovieapp.data.Actor
 
 class ActorsAdapter: RecyclerView.Adapter<ActorViewHolder>() {
