+++ app/src/main/java/ru/aevd/androidacademymovieapp/ui/adapters/OnMoviesItemClicked.kt
@@ -1,4 +1,4 @@
-package ru.aevd.androidacademymovieapp
+package ru.aevd.androidacademymovieapp.ui.adapters
 
 import ru.aevd.androidacademymovieapp.data.Movie
